requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
openai>=1.50.0
httpx>=0.27.0
PyMuPDF>=1.23.8
pandas>=2.1.4
openpyxl>=3.1.2
Pillow>=10.2.0
python-docx>=1.1.0
reportlab>=4.0.0
